refactor: intialization timeout into request animation frame callback
This delay value did well most of the time but there were still times when it got it wrong. Any higher than this, the flash of no state to state being set was very noticeable. There could be a better way to do this, but I haven't been able figure it out yet.